+++ src/lib/mina.ts
@@ -1,5 +1,3 @@
-// This is for an account where any of a list of public keys can update the state
-
 import { Circuit, Ledger, LedgerAccount } from '../snarky.js';
 import { Field, Bool } from './core.js';
 import { UInt32, UInt64 } from './int.js';
@@ -9,7 +7,6 @@ import {
   addMissingSignatures,
   FeePayerUnsigned,
   ZkappCommand,
-  zkappCommandToJson,
   AccountUpdate,
   ZkappStateLength,
   ZkappPublicInput,
@@ -37,6 +34,7 @@ export {
   LocalBlockchain,
   currentTransaction,
   CurrentTransaction,
+  Transaction,
   setActiveInstance,
   transaction,
   sender,
@@ -61,7 +59,7 @@ interface TransactionId {
   hash(): string | undefined;
 }
 
-interface Transaction {
+type Transaction = {
   /**
    * Transaction structure used to describe a state transition on the Mina blockchain.
    */
@@ -96,7 +94,17 @@ interface Transaction {
    * Sends the {@link Transaction} to the network.
    */
   send(): Promise<TransactionId>;
-}
+};
+
+const Transaction = {
+  fromJSON(json: Types.Json.ZkappCommand): Transaction {
+    let transaction = ZkappCommand.fromJSON(json);
+    return newTransaction(
+      transaction,
+      (activeInstance as { proofsEnabled?: boolean }).proofsEnabled
+    );
+  },
+};
 
 type Account = Fetch.Account;
 
@@ -270,6 +278,10 @@ function createTransaction(
   };
 
   currentTransaction.leave(transactionId);
+  return newTransaction(transaction, proofsEnabled);
+}
+
+function newTransaction(transaction: ZkappCommand, proofsEnabled?: boolean) {
   let self: Transaction = {
     transaction,
     sign(additionalKeys?: PrivateKey[]) {
@@ -284,7 +296,7 @@ function createTransaction(
       return proofs;
     },
     toJSON() {
-      let json = zkappCommandToJson(self.transaction);
+      let json = ZkappCommand.toJSON(self.transaction);
       return JSON.stringify(json);
     },
     toPretty() {
@@ -357,6 +369,7 @@ function LocalBlockchain({
   const actions: Record<string, any> = {};
 
   return {
+    proofsEnabled,
     accountCreationFee: () => UInt64.from(accountCreationFee),
     currentSlot() {
       return UInt32.from(
@@ -415,7 +428,7 @@ function LocalBlockchain({
       txn.sign();
 
       let commitments = Ledger.transactionCommitments(
-        JSON.stringify(zkappCommandToJson(txn.transaction))
+        JSON.stringify(ZkappCommand.toJSON(txn.transaction))
       );
 
       if (enforceTransactionLimits)
@@ -436,7 +449,7 @@ function LocalBlockchain({
         }
       }
 
-      let zkappCommandJson = zkappCommandToJson(txn.transaction);
+      let zkappCommandJson = ZkappCommand.toJSON(txn.transaction);
       try {
         ledger.applyJsonTransaction(
           JSON.stringify(zkappCommandJson),
@@ -1206,15 +1219,14 @@ function verifyTransactionLimits(accountUpdates: AccountUpdate[]) {
   n2 := signedPair
   n1 := signedSingle
   
-  10.26*np + 10.08*n2 + 9.14*n1 < 69.45
-
   formula used to calculate how expensive a zkapp transaction is
-  */
 
+  10.26*np + 10.08*n2 + 9.14*n1 < 69.45
+  */
   let totalTimeRequired =
-    proofCost * authTypes['proof'] +
-    signedPairCost * authTypes['signedPair'] +
-    signedSingleCost * authTypes['signedSingle'];
+    proofCost * authTypes.proof +
+    signedPairCost * authTypes.signedPair +
+    signedSingleCost * authTypes.signedSingle;
 
   let isWithinCostLimit = totalTimeRequired < costLimit;